corrigindo nullpointer

+++ form/service/src/main/java/org/opensingular/form/persistence/dao/AttachmentDao.java
@@ -36,7 +36,7 @@ public class AttachmentDao<T extends AttachmentEntity, C extends AttachmentConte
 
     @Inject
     private AttachmentContentDao<C> attachmentContentDao;
-    
+
     public AttachmentDao() {
         super((Class<T>) AttachmentEntity.class);
     }
@@ -50,16 +50,18 @@ public class AttachmentDao<T extends AttachmentEntity, C extends AttachmentConte
         return o;
     }
 
-    public T insert(InputStream is, long length, String name){
+    public T insert(InputStream is, long length, String name) {
         C content = attachmentContentDao.insert(is, length);
         return insert(createAttachment(content, name));
     }
 
     public void delete(Long id) {
-        T t = get(id);
-        Long codContent = t.getCodContent();
-        delete(t);
-        attachmentContentDao.delete(codContent);
+        final T t = get(id);
+        if (t != null) {
+            Long codContent = t.getCodContent();
+            delete(t);
+            attachmentContentDao.delete(codContent);
+        }
     }
 
     public List<T> list() {
@@ -68,9 +70,9 @@ public class AttachmentDao<T extends AttachmentEntity, C extends AttachmentConte
     }
 
     protected T createAttachment(C content, String name) {
-        
+
         T fileEntity = createInstance();
-        
+
         fileEntity.setCodContent(content.getCod());
         fileEntity.setHashSha1(content.getHashSha1());
         fileEntity.setSize(content.getSize());
